Rename Timeline component to Steps and implement step rendering logic with audio notifications
